Added Main Menu/Game Over/Controls screens
Added screens, just with buttons, no titles or text up yet, but the
functionality of getting from one screen to another is there. In
"Player" class, un-comment the "decreaseHealth()" line in the "act"
method to immediately go to the Game Over Screen when the game starts,
to view its functionality.